+++ lionengine-core/src/main/java/com/b3dgs/lionengine/core/InputDeviceDirectional.java
@@ -56,14 +56,14 @@ public interface InputDeviceDirectional
     /**
      * Get the horizontal direction.
      * 
-     * @return <code>-1</code> if left, <code>1</code> if right, 0 if none.
+     * @return <code>negative</code> if left, <code>positive</code> if right, 0 if none.
      */
-    int getHorizontalDirection();
+    double getHorizontalDirection();
 
     /**
      * Get the horizontal direction.
      * 
-     * @return <code>1</code> if up, <code>1</code> if down, 0 if none.
+     * @return <code>positive</code> if up, <code>negative</code> if down, 0 if none.
      */
-    int getVerticalDirection();
+    double getVerticalDirection();
 }
